Move Dashboard to Dashboard.tsx
